Update enemy script, collision, and player base collision
Added Cutoutcharacter as in the course

The enemy would hit the base upon touching its edge but only disappear
when reaching the center of the base. By making the collision boxes
smaller there's now less of a felt delay.

The script changes are tweaks for consistency and to match the tutorial.